secret_sharing_project/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14.0,<3.15
psycopg[c]>=3.1.8,<3.2
cryptography~=3.4.8